--- oss/errors.go
@@ -1,7 +1,0 @@
-// +build oss
-
-package oss
-
-import "errors"
-
-var ErrNotSupported = errors.New("Feature available only in Dgraph Enterprise Edition.")
